AG-10257 - Get all headings from framework markdoc context to show on side nav
